+++ src/main/java/com/minusServer/www/app/repository/BillRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.minusServer.www.app.model.Bill;
 import com.minusServer.www.app.model.User;
 
-public interface BillRepository extends CrudRepository<Bill, Integer>{
+public interface BillRepository extends CrudRepository<Bill, Long>{
 
 	public List<Bill> findByUser(User user);
 }
