+++ src/main/java/com/minusServer/www/app/repository/ItemRepository.java
@@ -3,6 +3,6 @@ package com.minusServer.www.app.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.minusServer.www.app.model.Item;
 
-public interface ItemRepository extends CrudRepository<Item, Integer>{
+public interface ItemRepository extends CrudRepository<Item, Long>{
 
 }
